+++ crates/storage/lib.rs
@@ -4,7 +4,7 @@ mod api;
 mod rlp;
 mod store;
 pub mod store_db;
-mod trie_db;
+pub mod trie_db;
 #[cfg(feature = "rocksdb")]
 mod utils;