+++ crates/storage/store.rs
@@ -1,8 +1,8 @@
-use crate::api::StoreEngine;
-use crate::error::StoreError;
 use crate::store_db::in_memory::Store as InMemoryStore;
 #[cfg(feature = "rocksdb")]
 use crate::store_db::rocksdb::Store as RocksDBStore;
+use crate::{api::StoreEngine, apply_prefix};
+use crate::{error::StoreError, trie_db::layering::TrieLayerCache};
 
 use ethereum_types::{Address, H256, U256};
 use ethrex_common::{
@@ -710,14 +710,77 @@ impl Store {
         block_number: BlockNumber,
         address: Address,
         storage_key: H256,
+        trie_layer_cache: Arc<TrieLayerCache>,
     ) -> Result<Option<U256>, StoreError> {
         match self.get_block_header(block_number)? {
-            Some(header) => self.get_storage_at_root(header.state_root, address, storage_key),
+            Some(header) => {
+                self.get_storage_at_root(header.state_root, address, storage_key, &trie_layer_cache)
+            }
+            None => Ok(None),
+        }
+    }
+
+    pub fn get_storage_at_old(
+        &self,
+        block_number: BlockNumber,
+        address: Address,
+        storage_key: H256,
+    ) -> Result<Option<U256>, StoreError> {
+        match self.get_block_header(block_number)? {
+            Some(header) => self.get_storage_at_root_old(header.state_root, address, storage_key),
             None => Ok(None),
         }
     }
 
     pub fn get_storage_at_root(
+        &self,
+        state_root: H256,
+        address: Address,
+        storage_key: H256,
+        trie_layer_cache: &Arc<TrieLayerCache>,
+    ) -> Result<Option<U256>, StoreError> {
+        let hashed_address = hash_address(&address);
+        let account_hash = H256::from_slice(&hashed_address);
+        let storage_root_needed = if self.engine.flatkeyvalue_computed(account_hash)? {
+            // We will use FKVs, we don't need the root
+            false
+        } else {
+            let Some(account) =
+                self.get_account_state_by_root(state_root, address, trie_layer_cache)?
+            else {
+                return Ok(None);
+            };
+            if account.storage_root == *EMPTY_TRIE_HASH {
+                return Ok(None);
+            }
+            true
+        };
+        let hashed_key = hash_key(&storage_key);
+        let prefixed_key = apply_prefix(Some(account_hash), Nibbles::from_bytes(&hashed_key));
+        // let's check the cache now
+        let value = trie_layer_cache.get(state_root, prefixed_key.as_ref());
+
+        let value = match value {
+            Some(val) => Some(val),
+            None => {
+                let storage_trie = self.open_direct_storage_trie(account_hash, *EMPTY_TRIE_HASH)?;
+                let storage_root = if storage_root_needed {
+                    let rlp = storage_trie.db().get(Nibbles::default())?;
+                    rlp.map(keccak).unwrap_or(*EMPTY_TRIE_HASH)
+                } else {
+                    *EMPTY_TRIE_HASH
+                };
+                let storage_trie = self.open_direct_storage_trie(account_hash, storage_root)?;
+                storage_trie.get(&hashed_key)?
+            }
+        };
+        value
+            .filter(|rlp| !rlp.is_empty())
+            .map(|rlp| U256::decode(&rlp).map_err(StoreError::RLPDecode))
+            .transpose()
+    }
+
+    pub fn get_storage_at_root_old(
         &self,
         state_root: H256,
         address: Address,
@@ -736,13 +799,16 @@ impl Store {
             let account = AccountState::decode(&encoded_account)?;
             account.storage_root
         };
-        let storage_trie = self.open_storage_trie(account_hash, storage_root, state_root)?;
+        // let's check the cache now
 
+        let storage_trie = self.open_storage_trie(account_hash, storage_root, state_root)?;
         let hashed_key = hash_key(&storage_key);
-        storage_trie
+        let value = storage_trie
             .get(&hashed_key)?
             .map(|rlp| U256::decode(&rlp).map_err(StoreError::RLPDecode))
-            .transpose()
+            .transpose()?;
+
+        Ok(value)
     }
 
     pub async fn set_chain_config(&mut self, chain_config: &ChainConfig) -> Result<(), StoreError> {
@@ -892,9 +958,28 @@ impl Store {
         &self,
         state_root: H256,
         address: Address,
+        trie_layer_cache: &Arc<TrieLayerCache>,
     ) -> Result<Option<AccountState>, StoreError> {
-        let state_trie = self.open_state_trie(state_root)?;
-        self.get_account_state_from_trie(&state_trie, address)
+        let hashed_address = hash_address(&address);
+        if let Some(rlp) =
+            trie_layer_cache.get(state_root, Nibbles::from_bytes(&hashed_address).as_ref())
+        {
+            if rlp.is_empty() {
+                return Ok(None);
+            }
+            return Ok(Some(AccountState::decode(&rlp)?));
+        }
+        let state_trie = self.open_direct_state_trie(*EMPTY_TRIE_HASH)?;
+        let root = state_trie
+            .db()
+            .get(Nibbles::default())?
+            .map(keccak)
+            .unwrap_or(*EMPTY_TRIE_HASH);
+        let state_trie = self.open_direct_state_trie(root)?;
+        state_trie
+            .get(&hashed_address)?
+            .map(|rlp| AccountState::decode(&rlp).map_err(StoreError::RLPDecode))
+            .transpose()
     }
 
     pub fn get_account_state_from_trie(
@@ -1364,6 +1449,10 @@ impl Store {
 
     pub async fn create_checkpoint(&self, path: impl AsRef<Path>) -> Result<(), StoreError> {
         self.engine.create_checkpoint(path.as_ref()).await
+    }
+
+    pub fn get_trie_layer_cache(&self) -> Result<Arc<TrieLayerCache>, StoreError> {
+        self.engine.get_trie_layer_cache()
     }
 }